+++ lib/vagrant/provisioners/chef.rb
@@ -150,7 +150,7 @@ module Vagrant
         def validate(env, errors)
           super
 
-          errors.add(I18n.t("vagrant.config.chef.vagrant_as_json_key")) if @json.has_key?(:vagrant)
+          errors.add(I18n.t("vagrant.config.chef.vagrant_as_json_key")) if json.has_key?(:vagrant)
         end
 
         def instance_variables_hash